Moved config to Application namespace
