+++ php/hireinsocial/tests/HireInSocial/Tests/Application/Integration/HireInSocialTestCase.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace HireInSocial\Tests\Application\Integration;
 
 use HireInSocial\Tests\Application\Context\DatabaseContext;
-use HireInSocial\Config;
 use HireInSocial\Tests\Application\Context\ThrottleContext;
+use HireInSocial\Application\Config;
 use PHPUnit\Framework\TestCase;
 use function HireInSocial\bootstrap;
 use function HireInSocial\dbal;
